Bot/requirements.txt
aiohttp==3.7.4.post0
astroid==2.5.6
async-timeout==3.0.1
asyncio==3.4.3
attrs==21.2.0
audioread==2.1.9
beautifulsoup4==4.9.3
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
colorama==0.4.4
discord==1.0.1
discord-py-slash-command==1.2.2
discord.py==1.7.2
flake8==3.9.2
google-api-core==1.29.0
google-api-python-client==2.7.0
google-auth==1.30.1
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
httplib2==0.19.1
idna==2.10
isort==5.8.0
lazy-object-proxy==1.6.0
lyricsgenius==3.0.1
mccabe==0.6.1
multidict==5.1.0
mysql-connector-python==8.0.25
packaging==20.9
protobuf==3.17.2
psutil==5.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
pydub==0.25.1
pyflakes==2.3.1
pylint==2.8.3
PyNaCl==1.4.0
pyparsing==2.4.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
python-dotenv==0.17.1
pytz==2021.1
pywin32==301
requests==2.25.1
rsa==4.7.2
six==1.16.0
soupsieve==2.2.1
toml==0.10.2
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.5
wrapt==1.12.1
yarl==1.6.3
youtube-dl==2021.5.16
